requirements.txt
fastapi
uvicorn[standard]
ultralytics
pillow
huggingface_hub
